Seperating common methods for secagg_crypter in `_secagg_utils` file
Unit test for secagg_utilities

create test for weighted averaging over quantized values

imroved test readibility

Adding additional test and different values of weights to test
