+++ fedbiomed/common/secagg/__init__.py
@@ -3,7 +3,7 @@
 
 
 from ._dh import DHKey, DHKeyAgreement
-from ._jls import JoyeLibert, quantize, reverse_quantize
+from ._jls import JoyeLibert
 from ._secagg_crypter import SecaggCrypter, EncryptedNumber
 from ._secagg_dummy_crypter import SecaggLomCrypter
 from ._lom import LOM, PRF
@@ -16,7 +16,5 @@ __all__ = [
     "PRF",
     "EncryptedNumber",
     "SecaggCrypter",
-    "quantize",
-    "reverse_quantize",
     "SecaggLomCrypter"
 ]
