+++ fedbiomed/common/utils/__init__.py
@@ -34,6 +34,10 @@ from ._secagg_utils import (
     matching_parties_servkey,
     matching_parties_biprime,
     matching_parties_dh,
+    quantize,
+    reverse_quantize,
+    multiply,
+    divide
 )
 
 from ._versions import (
@@ -47,7 +51,6 @@ __all__ = [
     # _utils
     "read_file",
     "get_class_source",
-    "import_class_object_from_file"
     "import_class_from_spec",
     "get_ipython_class_file",
     "get_method_spec",
@@ -55,11 +58,11 @@ __all__ = [
     "convert_iterator_to_list_of_python_floats",
     "compute_dot_product",
     # _config_utils
-    ROOT_DIR,
-    CONFIG_DIR,
-    VAR_DIR,
-    CACHE_DIR,
-    TMP_DIR,
+    "ROOT_DIR",
+    "CONFIG_DIR",
+    "VAR_DIR",
+    "CACHE_DIR",
+    "TMP_DIR",
     "get_component_config",
     "get_component_certificate_from_config",
     "get_all_existing_config_files",
@@ -69,6 +72,10 @@ __all__ = [
     "matching_parties_servkey",
     "matching_parties_biprime",
     "matching_parties_dh",
+    "quantize",
+    "multiply",
+    "divide",
+    "reverse_quantize",
     # _versions
     "raise_for_version_compatibility",
     "__default_version__",
